Updated Snake
Reworked the way in which the snake is represented.  Instead of
representing it as a series of two or more points at fixed positions on
the screen, the snake is now represented as one or more sections.
Whilst this approach is somehow less intuitive, it makes writing the
program much easier!!

At this stage, there are a ton of bugs remaining to be fixed though.